# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Agent-Trace Collector API
  version: 0.0.0-unreleased
  description: |
    Local, unauthenticated Collector API for Agent-Trace. The default server is
    loopback-only and must not be exposed directly to an untrusted network.
servers:
  - url: http://127.0.0.1:4319
    description: Default local Collector
security: []
...
paths:
...
  /runs:
    get:
      tags: [Runs]
      summary: List Runs
      description: Returns a bounded Dashboard read model unless legacy mode is requested.
      operationId: listRuns
      parameters:
        - $ref: "#/components/parameters/IncludeUntracked"
        - $ref: "#/components/parameters/RunPage"
        - $ref: "#/components/parameters/RunPageSize"
        - name: q
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string, enum: [all, running, success, error] }
        - name: source
          in: query
          schema: { type: string }
        - name: model
          in: query
          schema: { type: string }
        - name: project
          in: query
          schema: { type: string }
        - name: environment
          in: query
          schema: { type: string }
        - name: tag
          in: query
          schema: { type: string }
        - name: favorite
          in: query
          schema: { type: boolean }
        - name: startedAfter
          in: query
          schema: { type: string, format: date-time }
        - name: startedBefore
          in: query
          schema: { type: string, format: date-time }
        - name: minCostUsd
          in: query
          schema: { type: number, minimum: 0 }
        - name: maxCostUsd
          in: query
          schema: { type: number, minimum: 0 }
        - name: sort
          in: query
          schema: { type: string, enum: [startedAt, name, status, duration, tokens, cost] }
        - name: order
          in: query
          schema: { type: string, enum: [asc, desc], default: desc }
        - $ref: "#/components/parameters/Legacy"
      responses:
        "200":
          description: Paginated Run page, or an unbounded Run array in legacy mode
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/DashboardRunPage"
                  - type: array
                    items:
                      $ref: "#/components/schemas/Run"
    post:
      tags: [Runs]
      summary: Create a Run
      operationId: createRun
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRun"
      responses:
        "201":
          $ref: "#/components/responses/Ok"
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          description: Run id is tombstoned
    delete:
      tags: [Runs]
      summary: Delete multiple Runs
      operationId: deleteRuns
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ids]
              properties:
                ids:
                  type: array
                  minItems: 1
                  items:
                    type: string
                    minLength: 1
      responses:
        "200":
          description: Deletion result
          content:
            application/json:
              schema:
                type: object
                required: [ok, deleted]
                properties:
                  ok:
                    const: true
                  deleted:
                    type: integer
                    minimum: 0
        "400":
          $ref: "#/components/responses/ValidationError"
  /runs/{id}:
    get:
      tags: [Runs]
      summary: Get one Run
      operationId: getRun
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          description: Dashboard Run with aggregate metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Runs]
      summary: Update Run status and result
      operationId: updateRun
      parameters:
        - $ref: "#/components/parameters/RunId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRun"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
        "400":
          $ref: "#/components/responses/ValidationError"
    delete:
      tags: [Runs]
      summary: Delete one Run and its Events
      operationId: deleteRun
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
        "404":
          $ref: "#/components/responses/NotFound"
  /runs/{id}/organization:
    patch:
      tags: [Runs]
      summary: Update Run project metadata and annotations
      operationId: updateRunOrganization
      parameters:
        - $ref: "#/components/parameters/RunId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                project: { type: [string, "null"], maxLength: 120 }
                environment: { type: [string, "null"], maxLength: 120 }
                version: { type: [string, "null"], maxLength: 120 }
                tags:
                  type: array
                  maxItems: 20
                  items: { type: string, minLength: 1, maxLength: 64 }
                note: { type: [string, "null"], maxLength: 2000 }
                favorite: { type: boolean }
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "404": { $ref: "#/components/responses/NotFound" }
  /runs/{id}/events:
    get:
      tags: [Events]
      summary: List Events for a Run
      description: Returns a filtered, bounded read model unless legacy mode is requested.
      operationId: listRunEvents
      parameters:
        - $ref: "#/components/parameters/RunId"
        - name: visibility
          in: query
          schema:
            type: string
            enum: [display, hidden, all]
            default: display
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 100
        - name: q
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - $ref: "#/components/parameters/Legacy"
      responses:
        "200":
          description: Paginated Event page, or an unbounded Event array in legacy mode
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/DashboardEventPage"
                  - type: array
                    items:
                      $ref: "#/components/schemas/Event"
  /runs/{id}/export:
    get:
      tags: [Runs]
      summary: Download a metadata-redacted Run snapshot
      operationId: exportRedactedRun
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          description: Redacted JSON attachment with stable pseudonymous ids
          headers:
            Content-Disposition:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RedactedRunExport"
        "404":
          $ref: "#/components/responses/NotFound"
  /analytics/runs/compare:
    get:
      tags: [Analytics]
      summary: Compare two to five Runs
      operationId: compareRuns
      parameters:
        - name: ids
          in: query
          required: true
          description: Comma-separated Run ids in display order.
          schema: { type: string }
      responses:
        "200":
          description: Bounded Run comparison
          content:
            application/json:
              schema:
                type: object
                required: [runs]
                properties:
                  runs:
                    type: array
                    minItems: 0
                    maxItems: 5
                    items:
                      $ref: "#/components/schemas/DashboardRunMetric"
        "400":
          $ref: "#/components/responses/ValidationError"
  /analytics/runs/trends:
    get:
      tags: [Analytics]
      summary: Get continuous UTC daily Run trends
      operationId: getRunTrends
      parameters:
        - name: days
          in: query
          schema: { type: integer, minimum: 1, maximum: 90, default: 14 }
      responses:
        "200":
          description: Daily Run, success, duration, token, and cost aggregates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardRunTrends"
  /runs/{id}/insights:
    get:
      tags: [Events]
      summary: Analyze deterministic Trace insights
      operationId: getRunInsights
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          description: Insight list for the complete Run
          content:
            application/json:
              schema:
                type: object
                required: [insights]
                properties:
                  insights:
                    type: array
                    items: { type: object, additionalProperties: true }
  /runs/{id}/tombstone:
    delete:
      tags: [Maintenance]
      summary: Allow a deleted Run id to be ingested again
      operationId: restoreDeletedRun
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "404": { $ref: "#/components/responses/NotFound" }
  /events:
    post:
      tags: [Events]
      summary: Create an Event
      operationId: createEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEvent"
      responses:
        "201":
          $ref: "#/components/responses/Ok"
        "400":
          $ref: "#/components/responses/ValidationError"
  /integrations/codex/hook:
    post:
      tags: [Integrations]
      summary: Ingest a Codex Hook payload
      operationId: ingestCodexHook
      parameters:
        - $ref: "#/components/parameters/Surface"
        - $ref: "#/components/parameters/SurfaceSource"
      requestBody:
        $ref: "#/components/requestBodies/IntegrationPayload"
      responses:
        "202":
          $ref: "#/components/responses/IngestionResult"
  /integrations/claude-code/hook:
    post:
      tags: [Integrations]
      summary: Ingest a Claude Code Hook payload
      operationId: ingestClaudeCodeHook
      requestBody:
        $ref: "#/components/requestBodies/IntegrationPayload"
      responses:
        "202":
          $ref: "#/components/responses/IngestionResult"
  /integrations/codex/otel/v1/logs:
    post:
      tags: [Integrations]
      summary: Ingest Codex OTLP/HTTP JSON logs
      operationId: ingestCodexOtelLogs
      parameters:
        - $ref: "#/components/parameters/Surface"
        - $ref: "#/components/parameters/SurfaceSource"
      requestBody:
        $ref: "#/components/requestBodies/IntegrationPayload"
      responses:
        "202":
          $ref: "#/components/responses/IngestionResult"
  /v1/logs:
    post:
      tags: [Integrations]
      summary: Ingest OTLP/HTTP JSON logs using the desktop-compatible path
      operationId: ingestDesktopOtelLogs
      requestBody:
        $ref: "#/components/requestBodies/IntegrationPayload"
      responses:
        "202":
          $ref: "#/components/responses/IngestionResult"
  /integrations/usage-scan:
    post:
      tags: [Integrations, Usage]
      summary: Replace or merge a local Usage Snapshot
      operationId: ingestUsageScan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [source, rows]
              properties:
                source:
                  const: tokscale
                complete:
                  type: boolean
                explicitClients:
                  type: array
                  items:
                    type: string
                reconciledClients:
                  type: array
                  items:
                    type: string
                transcriptClients:
                  type: array
                  items:
                    type: string
                transcriptSessionIds:
                  type: array
                  items:
                    type: string
                scannedAt:
                  type: string
                  format: date-time
                rows:
                  type: array
                  items:
                    $ref: "#/components/schemas/UsageRow"
                diagnostics:
                  type: array
                  items:
                    type: object
                    additionalProperties: true
                transcripts:
                  type: array
                  items:
                    type: object
                    additionalProperties: true
      responses:
        "202":
          $ref: "#/components/responses/IngestionResult"
...
  /maintenance/tombstones:
    get:
      tags: [Maintenance]
      summary: List deleted Run tombstones
      operationId: listRunTombstones
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      responses:
        "200":
          description: Bounded tombstone list
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
  /maintenance/privacy:
    get:
      tags: [Maintenance]
      summary: Get pre-storage redaction settings
      operationId: getPrivacySettings
      responses:
        "200":
          description: Sensitive keys and replacement text
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PrivacySettings" }
    put:
      tags: [Maintenance]
      summary: Replace pre-storage redaction settings
      operationId: updatePrivacySettings
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PrivacySettings" }
      responses:
        "200":
          description: Persisted privacy settings
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PrivacySettings" }
        "400": { $ref: "#/components/responses/ValidationError" }
...
  /maintenance/prune:
    post:
      tags: [Maintenance]
      summary: Delete Runs older than a cutoff
      operationId: pruneRuns
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [before]
              properties:
                before: { type: string, format: date-time }
                statuses:
                  type: array
                  items: { type: string }
                keepTombstones: { type: boolean, default: true }
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "400": { $ref: "#/components/responses/ValidationError" }
...
  /v1/traces:
    post:
      tags: [Integrations]
      summary: Ingest OTLP JSON traces
      operationId: ingestOtlpTraces
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        "202": { description: Trace payload accepted }
  /integrations/otlp/v1/traces:
    post:
      tags: [Integrations]
      summary: Ingest OTLP JSON traces through the explicit integration path
      operationId: ingestIntegrationOtlpTraces
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        "202": { description: Trace payload accepted }
...
  /analytics/budgets/{id}:
    delete:
      tags: [Analytics]
      summary: Delete an analytics budget
      operationId: deleteAnalyticsBudget
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "404": { description: Budget not found }
...
components:
  parameters:
    RunId:
      name: id
      in: path
      required: true
      schema:
        type: string
    IncludeUntracked:
      name: includeUntracked
      in: query
      schema:
        type: boolean
        default: false
    RunPage:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    RunPageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    Legacy:
      name: legacy
      in: query
      description: Return the old unbounded array response.
      schema:
        type: boolean
        default: false
...
  responses:
    Ok:
      description: Operation succeeded
      content:
        application/json:
          schema:
            type: object
            required: [ok]
            properties:
              ok:
                const: true
...
  schemas:
...
    Run:
      allOf:
        - $ref: "#/components/schemas/CreateRun"
        - type: object
          required: [status, startedAt]
          properties:
            endedAt:
              type: string
              format: date-time
            output: {}
            error: {}
...
    DashboardRunPage:
      type: object
      required: [runs, pagination, summary]
      properties:
        runs:
          type: array
          items:
            $ref: "#/components/schemas/Run"
        pagination:
          $ref: "#/components/schemas/Pagination"
        summary:
          type: object
          additionalProperties: true
...